+++ src/composables/toggle.ts
@@ -1,0 +1,13 @@
+import { ref } from 'vue'
+
+export const useToggle = () => {
+  const viewMode = ref<'grid' | 'list'>('grid')
+  const toggleView = (mode: 'grid' | 'list') => {
+    viewMode.value = mode
+  }
+
+  return {
+    viewMode,
+    toggleView,
+  }
+}
